requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# keystone dependencies
pbr>=0.6,!=0.7,<1.0
WebOb>=1.2.3
eventlet>=0.16.1
greenlet>=0.3.2
netaddr>=0.7.12
PasteDeploy>=1.5.0
Paste
Routes>=1.12.3,!=2.0
six>=1.7.0
SQLAlchemy>=0.9.7,<=0.9.99
sqlalchemy-migrate>=0.9.1,!=0.9.2
passlib
iso8601>=0.1.9
python-keystoneclient>=1.0.0
keystonemiddleware>=1.0.0
oslo.concurrency>=1.4.1         # Apache-2.0
oslo.config>=1.6.0  # Apache-2.0
oslo.messaging>=1.4.0,!=1.5.0
oslo.db>=1.4.1  # Apache-2.0
oslo.i18n>=1.3.0  # Apache-2.0
oslo.middleware>=0.3.0                  # Apache-2.0
oslo.serialization>=1.2.0               # Apache-2.0
oslo.utils>=1.2.0                       # Apache-2.0
oauthlib>=0.6
dogpile.cache>=0.5.3
jsonschema>=2.0.0,<3.0.0
pycadf>=0.6.0
posix_ipc
python-keyczar
msgpack-python
